Better abstraction of FieldExt, reduce dependency on Halo2 directly

+++ remainder_prover/src/zkdt/structs.rs
@@ -112,7 +112,7 @@ impl<F: FieldExt> MleAble<F> for DecisionNode<F> {
 impl<F: FieldExt> From<Vec<bool>> for BinDecomp16Bit<F> {
     fn from(bits: Vec<bool>) -> Self {
         BinDecomp16Bit::<F> {
-            bits: bits.iter().map(|x| F::from(*x)).collect::<Vec<F>>().try_into().unwrap()
+            bits: bits.iter().map(|x| F::from(*x as u64)).collect::<Vec<F>>().try_into().unwrap()
         }
     }
 }
@@ -120,7 +120,7 @@ impl<F: FieldExt> From<Vec<bool>> for BinDecomp16Bit<F> {
 impl<F: FieldExt> From<Vec<bool>> for BinDecomp4Bit<F> {
     fn from(bits: Vec<bool>) -> Self {
         BinDecomp4Bit::<F> {
-            bits: bits.iter().map(|x| F::from(*x)).collect::<Vec<F>>().try_into().unwrap()
+            bits: bits.iter().map(|x| F::from(*x as u64)).collect::<Vec<F>>().try_into().unwrap()
         }
     }
 }
